organized some files
